validate and parse msg now is shared accross transport nodes
